Grapple
Grapple: provide a shoot and a cancel shoot methodes the script split the logique in three part, the shooting, the dragging and the returning. The shoot function take a destination and a boolean as parameter the boolean is needed to know if there is something hit by the grapple or not, the dragging part is only possible when a wall is hit. when the grapple is cancel, a boost is applied to the owner when the grapple is grabbing something. finally the grapple handle a _grapple_returned() to set up things for being reusable
Player: small refactor of the scene, add grapple handling using the grapple scene, update the movement system to follow grapple use
Lobby: update wall to simplify collision computation, removed yellow light -> withe column extrude debug lab
Satus Effect: updated: now the first tick is at the creation of the effect, add a way to clean the effect if needed, add simplification and safe guard for no propagation (the speed boost after grappling for example doesn't spread)
Speed Effect: applying to target the speed bonus or malus, using the _clean_effect function to restore the speed of the player at the end of the effect
Health Component: handle speed modification (muliplication)